Updated few files

+++ Spring-Backend/src/main/java/com/example/SpringBackend/database/repos/ResidenceRepo.java
@@ -4,6 +4,6 @@ import com.example.SpringBackend.database.entities.Residences;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResidenceRepo extends JpaRepository<Residences, Integer> {
+public interface ResidenceRepo extends JpaRepository<Residences, Long> {
 
 }
